feat: registration form processing
