+++ manta-trusted-setup/src/bin/groth16_phase2_prepare.rs
