+++ manta-trusted-setup/src/bin/process_registration.rs
@@ -18,16 +18,52 @@
 
 use clap::{Parser, Subcommand};
 use manta_trusted_setup::groth16::ceremony::{
-    config::ppot::{Config, Participant},
+    config::ppot::{extract_registry, Config, Record},
     CeremonyError,
 };
-use manta_util::{
-    Array,
-};
-use std::{collections::HashMap, path::PathBuf};
+use manta_util::serde::{Deserialize, Serialize};
+use std::fs::File;
 
-/// Registry type
-type Registry = HashMap<Array<u8, 32>, Participant>;
+// run with
+// cargo run --release --package manta-trusted-setup --all-features --bin process_registration -- process /Users/thomascnorton/Downloads/Trusted_Setup_Signups /Users/thomascnorton/Documents/Manta/manta-rs/manta-trusted-setup/data/test_registry.csv
+
+/// These are the headers in the .csv automatically generated by the
+/// registration form.
+const EXPECTED_HEADERS: [&str; 14]= [
+    "First up, what\'s your first name?", 
+        "What is your email address? ", 
+        "Okay {{field:c393dfe5f7faa4de}}, what your signature?", 
+        "What\'s your public key, {{field:c393dfe5f7faa4de}}?", 
+        "Finally, what\'s your Twitter Handle?  ", 
+        "Alright {{field:c393dfe5f7faa4de}}, why is privacy important to you?", 
+        "We want to reward participation with a POAP designed to commemorate this historical Web 3 achievement. If you would like to receive one please share your wallet address", 
+        "score", 
+        "Finally, what\'s your Twitter Handle?  ", 
+        "What\'s your public key, {{field:c393dfe5f7faa4de}}?", 
+        "What\'s your Discord ID, {{field:c393dfe5f7faa4de}}?", 
+        "What is your motivation for participating in the Trusted Setup, {{field:c393dfe5f7faa4de}}?",
+        "Submitted At", 
+        "Token"
+];
+
+/// Short versions of the above headers to allow
+/// rows of the .csv to be deserialized to [`RegistrationInfo`].
+const SHORT_HEADERS: [&str; 14] = [
+    "name",
+    "email",
+    "signature",
+    "verifying_key_null",
+    "twitter_null",
+    "why_privacy",
+    "wallet",
+    "score",
+    "twitter",
+    "verifying_key",
+    "discord",
+    "motivation",
+    "submission_time",
+    "submission_token",
+];
 
 /// Command
 #[derive(Debug, Subcommand)]
@@ -54,8 +90,18 @@ impl Arguments {
         match self.command {
             Command::Process {
                 raw_registry_path,
-                registry_path
-            } => todo!()
+                registry_path,
+            } => {
+                let file =
+                    File::open(raw_registry_path).expect("Unable to open file raw registry file");
+                extract_registry::<RegistrationInfo>(
+                    &file,
+                    registry_path.into(),
+                    EXPECTED_HEADERS.into(),
+                    SHORT_HEADERS.into(),
+                )
+                .expect("Registry processing failed.")
+            }
         };
         Ok(())
     }
@@ -70,7 +116,112 @@ async fn main() {
         .expect("Server error occurred");
 }
 
-// run with
-// cargo run --release --all-features --bin groth16_phase2_server recover manta-trusted-setup/data manta-trusted-setup/data/registry.csv
-// TODO: Update server start command!
-// cargo build --release --all-features --bin groth16_phase2_server
+/// Registration info collected by our registration form.
+#[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
+#[serde(
+    bound(deserialize = "", serialize = ""),
+    crate = "manta_util::serde",
+    deny_unknown_fields
+)]
+pub struct RegistrationInfo {
+    /// First name (may be empty)
+    pub name: String,
+
+    /// Email Account
+    pub email: String,
+
+    /// Signature
+    pub signature: String,
+
+    /// Verifying Key
+    pub verifying_key: String,
+
+    /// Twitter Account
+    pub twitter: String,
+
+    /// Why is privacy important (may be empty)
+    pub why_privacy: String,
+
+    /// Wallet address (may be empty)
+    pub wallet: String,
+
+    /// Score
+    pub score: String,
+
+    /// Named field that's always empty
+    pub twitter_null: String,
+
+    /// Named field that's always empty
+    pub verifying_key_null: String,
+
+    /// Discord ID (may be empty)
+    pub discord: String,
+
+    /// Motivation to participate (may be empty)
+    pub motivation: String,
+
+    /// Submission time
+    pub submission_time: String,
+
+    /// Submission token
+    pub submission_token: String,
+}
+
+impl From<RegistrationInfo> for Record {
+    fn from(value: RegistrationInfo) -> Self {
+        Self::new(
+            value.twitter,
+            value.email,
+            "false".to_string(),
+            value.verifying_key,
+            value.signature,
+        )
+    }
+}
+
+#[test]
+fn test_extract_registry() {
+    use manta_trusted_setup::groth16::ceremony::config::ppot::extract_registry;
+    let file =
+        File::open("/Users/thomascnorton/Downloads/Trusted Setup Signups - Trusted Setup (2).csv")
+            .expect("Cannot open file");
+    let path = PathBuf::from(
+        r"/Users/thomascnorton/Documents/Manta/manta-rs/manta-trusted-setup/data/test_registry.csv",
+    );
+    extract_registry::<RegistrationInfo>(
+        &file,
+        path,
+        EXPECTED_HEADERS.into(),
+        SHORT_HEADERS.into(),
+    )
+    .unwrap();
+}
+
+#[test]
+fn test_set_headers() {
+    use csv::Reader;
+    use manta_trusted_setup::groth16::ceremony::config::ppot::set_header;
+
+    let file = File::open("/Users/thomascnorton/Documents/Manta/manta-rs/manta-trusted-setup/data/registry_buffer.csv").expect("Cannot open file");
+    let mut reader = Reader::from_reader(&file);
+    assert!(set_header(&mut reader, EXPECTED_HEADERS.into(), SHORT_HEADERS.into()).is_ok());
+    assert!(
+        reader.byte_headers().unwrap()
+            == vec![
+                "name",
+                "email",
+                "signature",
+                "verifying_key_null",
+                "twitter_null",
+                "why_privacy",
+                "wallet",
+                "score",
+                "twitter",
+                "verifying_key",
+                "discord",
+                "motivation",
+                "submission_time",
+                "submission_token",
+            ]
+    );
+}
